% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernsteinSpl.R
\name{bernsteinSpline}
\alias{bernsteinSpline}
\title{Bernstein Polynomial Basis}
\usage{
bernsteinSpline(x, df, ...)
}
\arguments{
\item{x}{The predictor variable. Missing values are allowed.}

\item{df}{Degrees of freedom.
## @param Boundary.knots Boundary points at wich to anchor the Bernstein
#basis.}

\item{...}{Optional arguments for future usage.}
}
\value{
A matrix of dimension \code{length(x)} by \code{df}.
TODO: Attributes of derivatives, integral, etc.
}
\description{
Generate the Bernstein spline basis matrix given a degree of freedom
}
\examples{
## examples

}
\references{
\url{https://people.sc.fsu.edu/~jburkardt/c_src/bernstein_polynomial/bernstein_polynomial.html}
}
