% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv.R
\name{deriv}
\alias{deriv}
\alias{deriv.bSpline2}
\alias{deriv.cSpline}
\alias{deriv.iSpline}
\alias{deriv.ibs}
\title{Derivative of Splines}
\usage{
\method{deriv}{ibs}(expr, derivs = 1L, ...)

\method{deriv}{iSpline}(expr, derivs = 1L, ...)

\method{deriv}{cSpline}(expr, derivs = 1L, ...)

\method{deriv}{bSpline2}(expr, derivs = 1L, ...)
}
\arguments{
\item{expr}{\code{spline2} objects generated from this package.}

\item{derivs}{A positive integer indicating the number of
derivatives.}

\item{...}{Other arguments for further usage.}
}
\description{
Obtain derivative of spline bases.
}

