% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.bSpline2}
\alias{print.ibs}
\alias{print.dbs}
\alias{print.mSpline}
\alias{print.iSpline}
\alias{print.cSpline}
\alias{print.deriv}
\title{Print Out a Spline Basis Matrix}
\usage{
\method{print}{bSpline2}(x, ...)

\method{print}{ibs}(x, ...)

\method{print}{dbs}(x, ...)

\method{print}{mSpline}(x, ...)

\method{print}{iSpline}(x, ...)

\method{print}{cSpline}(x, ...)

\method{print}{deriv}(x, ...)
}
\arguments{
\item{x}{Objects of class \code{bSpline2}, \code{ibs}, \code{mSpline},
\code{iSpline}, or \code{cSpline}, etc.}

\item{...}{Optional argument for future usage.}
}
\value{
Object input.
}
\description{
\code{Print} methods that simply print out the spline basis matrix without
unnecessary attributes.
}
