% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSpline.R
\name{cSpline}
\alias{cSpline}
\title{C-Spline Basis for Polynomial Splines}
\usage{
cSpline(x, df = NULL, knots = NULL, degree = 3, intercept = FALSE,
        Boundary.knots = range(x), rescale = TRUE, ...)
}
\arguments{
\item{x}{The predictor variable.  Missing values are allowed and will be
returned as they were.}

\item{df}{Degrees of freedom.  One can specify \code{df} rather than
\code{knots}, then the function chooses "df - degree"
(minus one if there is an intercept) knots at suitable quantiles of \code{x}
(which will ignore missing values).  The default, \code{NULL}, corresponds
to no inner knots, i.e., "degree - intercept".}

\item{knots}{The internal breakpoints that define the spline.  The
default is \code{NULL}, which results in a basis for ordinary
polynomial regression.  Typical values are the mean or median
for one knot, quantiles for more knots.  See also
\code{Boundary.knots}.}

\item{degree}{Non-negative integer degree of the piecewise polynomial. The
default value is 3 for cubic splines.}

\item{intercept}{If \code{TRUE}, an intercept is included in the basis;
Default is \code{FALSE}.}

\item{Boundary.knots}{Boundary points at which to anchor the C-spline basis.
By default, they are the range of the non-\code{NA} data.  If both
\code{knots} and \code{Boundary.knots} are supplied, the basis parameters
do not depend on \code{x}. Data can extend beyond \code{Boundary.knots}.}

\item{rescale}{Logical value (\code{TRUE} by default) indicating whether
rescaling on C-spline basis is required. If TRUE, C-spline basis is rescaled
to have unit height at right boundary knot; the corresponding I-spline and
M-spline basis matrices shipped in attributes are also rescaled to the same
extent.}

\item{...}{Optional arguments for future usage.}
}
\value{
A matrix of dimension \code{length(x)} by
\code{df = degree + length(knots)} (plus on if intercept is included).
Attributes that correspond to the arguments specified are returned
for usage for \code{\link{predict.cSpline}}. The corresponding M-spline and
I-spline basis matrices are also returned in attribute named \code{msMat}
and \code{isMat} respectively.
}
\description{
This function generates the convex regression spline (called C-spline) basis
matrix by integrating I-spline basis for a polynomial spline.
}
\details{
It is an implementation of the close form C-spline basis derived from
the recursion formula of I-spline and M-spline.  Internally, it calls
\code{\link{iSpline}} and generates a basis matrix for representing the
family of piecewise polynomials and their corresponding integrals with the
specified interior knots and degree, evaluated at the values of \code{x}.
}
\examples{
library(graphics)
x <- seq(0, 1, by = .01)
knots <- c(0.3, 0.5, 0.6)
cMat <- cSpline(x, knots = knots, degree = 2, intercept = TRUE)
matplot(x, cMat, type = "l", ylab = "C-spline basis")
abline(v = knots, lty = 2, col = "gray")
matplot(x, attr(cMat, "isMat"), type = "l", ylab = "rescaled I-spline basis")
matplot(x, attr(cMat, "msMat"), type = "l", ylab = "rescaled M-spline basis")
}
\references{
Meyer, M. C. (2008). Inference using shape-restricted regression splines.
\emph{The Annals of Applied Statistics}, 1013--1033. Chicago
}
\seealso{
\code{\link{predict.cSpline}} for evaluation at given (new) values;
\code{\link{iSpline}} for I-spline basis;
\code{\link{mSpline}} for M-spline basis;
\code{\link{bSpline}} for B-spline basis;
\code{\link{ibs}} for integral of B-spline basis.
}

