% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.bSpline2}
\alias{print.cSpline}
\alias{print.iSpline}
\alias{print.ibs}
\alias{print.mSpline}
\title{Print out spline basis matrix}
\usage{
\method{print}{bSpline2}(x, ...)

\method{print}{ibs}(x, ...)

\method{print}{mSpline}(x, ...)

\method{print}{iSpline}(x, ...)

\method{print}{cSpline}(x, ...)
}
\arguments{
\item{x}{Objects of class \code{bSpline2}, \code{ibs}, \code{mSpline},
\code{iSpline}, or \code{cSpline}, etc.}

\item{...}{Optional argument for future usage.}
}
\value{
The object inputted.
}
\description{
This function simply prints out the spline basis matrix without unnecessary
attributes and invisibly returns the object inputted.
}

