% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines-package.R
\docType{package}
\name{splines2}
\alias{splines2}
\alias{splines2-package}
\title{splines2: Regression Spline Functions and Classes Too}
\description{
A complementary package on splines providing functions
constructing M-spline, I-spline, and integral of B-spline basis.
}

