% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines2-package.R
\docType{package}
\name{splines2}
\alias{splines2}
\alias{splines2-package}
\title{splines2: Regression Spline Functions and Classes Too}
\description{
A complementary package on splines providing functions
constructing B-splines, integral of B-splines, monotone splines
(M-splines) and its integral (I-splines), convex splines (C-splines),
and their derivatives of given order. Piecewise constant basis is
allowed for B-spline and M-spline basis.
}

