% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines2-package.R
\docType{package}
\name{splines2}
\alias{splines2}
\alias{splines2-package}
\title{splines2: Regression Spline Functions and Classes}
\description{
A supplementary package on splines providing functions constructing
B-splines, integral of B-splines, monotone splines (M-splines) and its
integral (I-splines), convex splines (C-splines), and their derivatives of
given order. Piecewise constant basis is allowed for B-spline and M-spline
basis.
}
\details{
It is named after the package \pkg{splines}: ``Regression Spline Functions
and Classes''. The tailing number two is simply ``too'' (and by no means for
the generation two).
}
