% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines2-package.R
\docType{package}
\name{splines2}
\alias{splines2}
\title{splines2: Regression Spline Functions and Classes}
\description{
This package provides functions to construct basis matrix of
\itemize{
\item B-splines
\item M-splines
\item I-splines
\item convex splines (C-splines)
\item generalized Bernstein polynomials
\item their integrals (except C-splines) and derivatives of given order by
close-form recursive formulas
}
}
\details{
In addition to the R interface, it also provides a C++ header-only library
integrated with \pkg{Rcpp}, which allows construction of spline basis matrix
directly in C++ with the help of \pkg{Rcpp} and \pkg{RcppArmadillo}.  So it
can also be treated as one of the \pkg{Rcpp*} packages.  A toy example
package that uses the C++ interface is available at
<https://github.com/wenjie2wang/example-pkg-Rcpp-splines2>.

It is named after the \pkg{splines} package:
"Regression Spline Functions and Classes". The tailing number two is simply
"too" (and by no means for the generation two).
}
