% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines2-package.R
\docType{package}
\name{splines2}
\alias{splines2}
\alias{splines2-package}
\title{splines2: Regression Spline Functions and Classes Too}
\description{
A complementary package on splines providing functions constructing
B-spline, monotone spline (M-spline) and its integral (I-spline), convex
spline (C-spline), and integral of B-spline basis.  Piecewise constant basis
is allowed for B-spline and M-spline basis.
}

